feat(MainPage): basic refactored scripts with custom hook
